[ADH-4646] Update the way of supporting multiple HDFS versions
