+++ smart-engine/pom.xml
@@ -132,6 +132,16 @@
             <artifactId>sqlite-jdbc</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.smartdata</groupId>
+            <artifactId>${smart.hadoop.artifact}</artifactId>
+            <version>1.6.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.smartdata</groupId>
+            <artifactId>${smart.hadoop.client.artifact}</artifactId>
+            <version>1.6.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>org.dbunit</groupId>
             <artifactId>dbunit</artifactId>
             <scope>test</scope>
@@ -175,56 +185,6 @@
     </dependencies>
 
     <profiles>
-        <profile>
-            <id>hadoop-3.1</id>
-            <dependencies>
-                <dependency>
-                    <groupId>org.smartdata</groupId>
-                    <artifactId>smart-hadoop-3.1</artifactId>
-                    <version>1.6.0-SNAPSHOT</version>
-                </dependency>
-                <dependency>
-                    <groupId>org.smartdata</groupId>
-                    <artifactId>smart-hadoop-client-3.1</artifactId>
-                    <version>1.6.0-SNAPSHOT</version>
-                </dependency>
-            </dependencies>
-        </profile>
-        <profile>
-            <id>hadoop-3.2</id>
-            <dependencies>
-                <dependency>
-                    <groupId>org.smartdata</groupId>
-                    <artifactId>smart-hadoop-3.2</artifactId>
-                    <version>1.6.0-SNAPSHOT</version>
-                    <exclusions>
-                        <exclusion>
-                            <groupId>ch.qos.reload4j</groupId>
-                            <artifactId>reload4j</artifactId>
-                        </exclusion>
-                        <exclusion>
-                            <groupId>org.slf4j</groupId>
-                            <artifactId>slf4j-reload4j</artifactId>
-                        </exclusion>
-                    </exclusions>
-                </dependency>
-                <dependency>
-                    <groupId>org.smartdata</groupId>
-                    <artifactId>smart-hadoop-client-3.2</artifactId>
-                    <version>1.6.0-SNAPSHOT</version>
-                    <exclusions>
-                        <exclusion>
-                            <groupId>ch.qos.reload4j</groupId>
-                            <artifactId>reload4j</artifactId>
-                        </exclusion>
-                        <exclusion>
-                            <groupId>org.slf4j</groupId>
-                            <artifactId>slf4j-reload4j</artifactId>
-                        </exclusion>
-                    </exclusions>
-                </dependency>
-            </dependencies>
-        </profile>
         <profile>
             <id>alluxio</id>
             <dependencies>
