+++ smart-server/pom.xml
@@ -199,6 +199,24 @@
             <version>1.7.16</version>
         </dependency>
         <dependency>
+            <groupId>org.eclipse.jetty</groupId>
+            <artifactId>jetty-http</artifactId>
+            <version>${jetty.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.smartdata</groupId>
+            <artifactId>${smart.hadoop.artifact}</artifactId>
+            <version>1.6.0-SNAPSHOT</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.smartdata</groupId>
+            <artifactId>${smart.hadoop.artifact}</artifactId>
+            <version>1.6.0-SNAPSHOT</version>
+            <scope>test</scope>
+            <type>test-jar</type>
+        </dependency>
+        <dependency>
             <groupId>org.smartdata</groupId>
             <artifactId>smart-metastore</artifactId>
             <version>1.6.0-SNAPSHOT</version>
@@ -221,55 +239,6 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
-
-    <profiles>
-        <profile>
-            <id>hadoop-3.1</id>
-            <dependencies>
-                <dependency>
-                    <groupId>org.smartdata</groupId>
-                    <artifactId>smart-hadoop-3.1</artifactId>
-                    <version>1.6.0-SNAPSHOT</version>
-                    <scope>test</scope>
-                </dependency>
-                <dependency>
-                    <groupId>org.smartdata</groupId>
-                    <artifactId>smart-hadoop-3.1</artifactId>
-                    <version>1.6.0-SNAPSHOT</version>
-                    <scope>test</scope>
-                    <type>test-jar</type>
-                </dependency>
-                <dependency>
-                    <groupId>org.eclipse.jetty</groupId>
-                    <artifactId>jetty-http</artifactId>
-                    <version>${jetty.version}</version>
-                </dependency>
-            </dependencies>
-        </profile>
-        <profile>
-            <id>hadoop-3.2</id>
-            <dependencies>
-                <dependency>
-                    <groupId>org.smartdata</groupId>
-                    <artifactId>smart-hadoop-3.2</artifactId>
-                    <version>1.6.0-SNAPSHOT</version>
-                    <scope>test</scope>
-                </dependency>
-                <dependency>
-                    <groupId>org.smartdata</groupId>
-                    <artifactId>smart-hadoop-3.2</artifactId>
-                    <version>1.6.0-SNAPSHOT</version>
-                    <scope>test</scope>
-                    <type>test-jar</type>
-                </dependency>
-                <dependency>
-                    <groupId>org.eclipse.jetty</groupId>
-                    <artifactId>jetty-http</artifactId>
-                    <version>${jetty.version}</version>
-                </dependency>
-            </dependencies>
-        </profile>
-    </profiles>
 
     <build>
         <plugins>
